requirements.txt
numpy>=1.20.0
matplotlib>=3.3.0
scipy>=1.6.0
numba>=0.56.0
cupy-cuda12x>=12.0.0; sys_platform != "darwin"  # CUDA 12.x support, optional
cupy-cuda11x>=11.0.0; sys_platform != "darwin"  # CUDA 11.x support, optional
